Webpage Finished.
All the comments were added not only in the CSS page, but the index as well because of how it was created from scratch. In addition to that, I also added a cougar image near the appropriate story and made a couple tweaks from feedback on how to make it look better.