Additional Touch-UPs
In response to me realizing today that the project was due on Wednesday and not Monday, I decided to make a couple more changes I thought of after initially turning the project in. These include emphasized words in the stories, a page title and image, and an alteration to the background color. Comments were also added in alongside these updates.